Split up actions into files
