+++ src/components/Reviewer/sentence.js
@@ -3,7 +3,12 @@ import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import { useDispatch, useSelector } from 'react-redux';
 import { useState, useEffect } from 'react';
-import { fetchSentenceToBeReviewed, doneFetchingSentence, submitReview } from '../../redux/actions';
+
+import { 
+    doneFetchingSentence, fetchSentence, fetchSentenceToBeReviewed
+    } from '../../redux/actions/sentenceActions';
+import { addPeopleReview } from '../../redux/actions/reviewActions';
+
 import { store } from '../../redux/store'
 import Popover from '@material-ui/core/Popover';
 import PopupState, { bindTrigger, bindPopover } from 'material-ui-popup-state';
@@ -57,7 +62,7 @@ const Sentence = () => {
     const classes = useStyles();
     const dispatch = useDispatch();
     const state = useSelector(state => state);
-    const {isFetchingSentence} = state;
+    const { isFetchingSentence } = state;
 
     useEffect(() => {
         new Promise((resolve, reject) => {
@@ -80,7 +85,14 @@ const Sentence = () => {
     ? <CircularProgress color="secondary"/>
     : null;
 
-    const { sentence, user, rules } = state; 
+    // submitReview({
+    //     ...review,
+    //     ruleReview: 1, // 1 means correct
+    // })
+
+    const { rules } = state.ruleReducer;
+    const { user } = state.userReducer;
+    const { sentence } = state.sentenceReducer; 
 
     let review = {}
     if (sentence !== undefined) {
@@ -89,7 +101,7 @@ const Sentence = () => {
             reviewerID: user.id,
             dateAdded: "now()"
         }
-    } 
+    }
 
     return (
         <div className={`${classes.root} ${classes.GridItem}`}>
@@ -114,10 +126,7 @@ const Sentence = () => {
             </CardContent>
             <CardActions className={classes.buttons}>
                 <Button variant="contained" size="medium" color="secondary" onClick={() =>
-                    dispatch(submitReview({
-                        ...review,
-                        ruleReview: 1, // 1 means correct
-                    }))
+                    dispatch()
                 } className={classes.buttons.correct}>Correct</Button>
                 <PopupState variant="popover" popupId="demo-popup-popover">
                     {popupState => (
