+++ src/redux/reducers/users.js
@@ -2,9 +2,20 @@
 // Daniel Nicolas Gisolfi
 
 import * as ActionTypes from '../constants';
-import { initialState } from './index';
 
-export const userReducer = (state = initialState, action) => {
+
+export const defaultState = {
+	loggedIn: false,
+	user: {
+        first: 'Daniel',
+        last: 'Gisolfi',
+        email: 'Daniel.Gisolfi1@marist.edu',
+        id: 20074558
+	},
+	error: null
+}
+
+export const userReducer = (state = defaultState, action) => {
 	switch (action.type) {
 
 		default:
